string : check if malloc failed on strdup/strndup

+++ libc/string/strnlen.c
@@ -3,5 +3,5 @@
 
 size_t strnlen(const char *str, size_t maxlen) {
 	const char *end = memchr(str, '\0', maxlen);
-	return end ? (uintptr_t)end - (uintptr_t)str : maxlen;
+	return end ? (size_t)(end - str) : maxlen;
 }
